Add unit tests for home price appreciation calculator

+++ RentOrBuy.Home.DataModel/OwnershipCost/OwnershipCostFactors.cs
@@ -10,7 +10,7 @@ namespace RentOrBuy.Home.DataModel.OwnershipCost
     public record OwnershipCostFactors
     {
         public uint Price { get; init; }
-        public ushort PlannedLengthOfStay { get; set; }  
+        public byte PlannedLengthOfStay { get; set; }  
         [Description("Projected annual home price growth as percentage of home value")]
         public decimal AnnualPriceGrowthRate { get; set; }
